Allow runners to intercept subprocess execution
Instead of directly executing the runner subprocess in the `run` method
of the `BaseRunner`, rather call a new method `run_subprocess` which
deals with the subprocess spinup. This allows subclasses to override the
latter method in order to avoid executing a subprocess altogether, and
inject a custom log output and return code. Runner-specific blacklists
can be implemented in this manner.

Signed-off-by: Fabian Schuiki <fabian@schuiki.ch>