Add equality check for escaped identifiers
Following the description of Chapter 5.6.1 the backslash or following
white space is not part of the identifier.
As an example from the specification the escaped identifier `\cpu3` must
the same as the nonescaped identifier `cpu3`.

Add an escaped identifier and use it nonescaped.
Escalate the undefined usage to an error.

Signed-off-by: Tobias Wölfel <tobias.woelfel@mailbox.org>