Normalize CSV output and emit relevant data for each (test, tool) tuple.
Previous CSV file had one column per tool and in each of these only the
outcome of the test: True or False.

Now, have one row per (test, tool) tuple and output more relevant
information.

  * `TestName`, `Tool`  *the selector*
  * `Pass`  *{True or False}*
  * `Tags`  *associated tags with the TestName; space separated*
  * `InputBytes` *Bytes the input (System)Verilog files add up to*
  * `ExitCode` *Exit code of the tool. Special value 71 for timeout (`71meout` :))*
  * `AllowedTimeout`  *Timeout allowed for this test*
  * `TimeUser`, `TimeSystem`, `TimeWall`
  * `RamUsageMiB`

We are not emitting all associated filenames anymore per test - this was partially
a huge list for some of the tests, and not terribly useful. The TestName is unique.

Also: file content of report.csv changed, so someone who was relying on the old format
is breaking. However I'd say someone who is well-versed enough to use the CSV, will
quickly fix things, so we should not have an ugly backward compatible version.

Signed-off-by: Henner Zeller <h.zeller@acm.org>