Improve github cache handling
Github caches tend to get buggy with GCP runners that we use here.
If they work correctly, the caches are set up in ~45s max.
If they don't, this steps tends to take over 360 minutes (the default
timeout) and break the CI.

This patch tries to workaround this problem by setting the timeout to 3
minutes but making the cache step optional. If we fail to set up caches,
just build everything normally, it will be quicker anyway.

Signed-off-by: Tomasz Gorochowik <tgorochowik@antmicro.com>