history graph: properly interpret date in data points
There are days where there are fewer or more builds that 1 a day. Each
build generates a new entry in the history graph.

Without this change, each entry created a new evenly separated data
point which didn't really make sense when analyzing the history.

With this change the data points will be split in accordance to the
actual build timestamp.

As a side benefit we have a nicer formatting of the dates.
Full timestamps can still be checked by hovering individual data points.

Signed-off-by: Tomasz Gorochowik <tgorochowik@antmicro.com>