runners: icarus: Fix top-level module specification
Currently when specifying a top-level module for the icarus runner the `-s`
switch is passed together with the top-level module name as a single
parameter. Both are separated by a space, e.g. `-s top`. This results in the
space being included in the top-level module name, resulting in errors such
as

    error: Unable to find the root module " top" in the Verilog source.

Pass to the `-s` switch and the top-level module name as two separate
parameters to fix this.

Signed-off-by: Lars-Peter Clausen <lars@metafoo.de>