Fix wildcard equality operator tests with x/z in LHS
For the wildcard equality operator only x and z bits of the right hand
side value are considered as wildcard bits.

x and z bits on the left hand side are still compared and may result in the
comparison itself returning x.

There are a few test entries for the wildcard equality operator test that
do not consider this and expect the wrong value. Correct these tests to
expect that the result is 1'bx.

Signed-off-by: Lars-Peter Clausen <lars@metafoo.de>