Temporarily use older ubuntu image
Latest points to jammy, this does not change the distro, but takes an
older Jammy release.

The new one seems to not be compatible with Singularity (which we're
using in our custom runners).

This change can be reverted once the Singularity issue is resolved.
Alternatively we can switch to other distro (debian seems to work).

Signed-off-by: Tomasz Gorochowik <tgorochowik@antmicro.com>