Avoid race conditions between `assign` and `initial`
The initial execution of a `assign` statement is scheduled in the same time
slot as an `initial` block.

There are a few tests that use an `initial` block to evaluate whether an
`assign` statement was correctly executed. But this leads to an inherit
race condition as the `initial` block might run before the `assign` has
been executed.

Switch these tests to use a `final` block instead to avoid the race
condition.

Signed-off-by: Lars-Peter Clausen <lars@metafoo.de>