[Tools] Add the new tool - circt-verilog (#5866)
The circt-verilog is a hardware compiler based on circt & slang frontend framework. It provides
ultimate possibility to let SV step in the world of MLIR. It is still in early stage of developing.
Testing results coming from sv-tests is necessary for us to support and optimize the circt-verilog
tool. To promote more advanced features and fixes into the tool. It is the initial commit for it,
partial configurations could refer to the config of slang compiler.